refactor(core): CATALYST-00 create _components directory in product
