+++ apps/core/app/(default)/product/[slug]/_components/Breadcrumbs.tsx
