+++ apps/core/app/(default)/product/[slug]/_components/Gallery.tsx
