+++ apps/core/app/(default)/product/[slug]/_components/ProductReviewSchema.tsx
