+++ apps/core/app/(default)/product/[slug]/_components/ProductSchema.tsx
