+++ apps/core/app/(default)/product/[slug]/_components/RelatedProducts.tsx
