+++ apps/core/app/(default)/product/[slug]/_components/ReviewSummary.tsx
